refactor library structure
