+++ src/protocols/chainsync_protocol.rs
@@ -22,7 +22,7 @@ use rusqlite::{Connection, Error, named_params, NO_PARAMS};
 use serde::Serialize;
 use serde_cbor::{de, ser, Value};
 
-use super::{Agency, Protocol};
+use crate::{Agency, Protocol};
 use super::chainsync_protocol::msg_roll_backward::parse_msg_roll_backward;
 use super::chainsync_protocol::msg_roll_forward::{MsgRollForward, parse_msg_roll_forward, Tip};
